+++ internal/quickswitch/go.mod
@@ -1,8 +1,5 @@
-module github.com/HellstromIT/go-quickswitch/internal/cli
+module github.com/HellstromIT/go-quickswitch/internal/quickswitch
 
 go 1.15
 
-require (
-	github.com/alecthomas/kong v0.2.11
-	github.com/ktr0731/go-fuzzyfinder v0.2.1
-)
+require github.com/alecthomas/kong v0.2.11
