+++ pkgs/fuzzy/go.mod
@@ -1,0 +1,5 @@
+module github.com/HellstromIT/go-quickswitch/pkgs/fuzzy
+
+go 1.15
+
+require github.com/ktr0731/go-fuzzyfinder v0.2.1
